rename model to deployment
