+++ packages/client/src/ee/main.tsx
@@ -38,7 +38,7 @@ import AppDetail from 'containers/appDetail';
 import AppEdit from 'containers/appEditPage';
 import GroupSettingsPage from 'containers/groupSettingsPage';
 import GroupSettingsJobs from 'ee/components/groupSettings/jobs';
-import GroupSettingsModels from 'ee/components/groupSettings/models';
+import GroupSettingsDeployments from 'ee/components/groupSettings/deployments';
 import GroupSettingsMLflow from 'ee/components/groupSettings/mlflow';
 
 const client = createGraphqlClient({
@@ -84,9 +84,9 @@ class Main extends React.Component {
                 extraTabs={[
                   { component: GroupSettingsJobs, key: 'jobs', tab: 'Jobs' },
                   {
-                    component: GroupSettingsModels,
-                    key: 'models',
-                    tab: 'Models',
+                    component: GroupSettingsDeployments,
+                    key: 'deployments',
+                    tab: 'Deployments',
                   },
                   {
                     component: GroupSettingsMLflow,
